pyproject.toml
[project]
name = "ebay-searcher"
version = "0.1.0"
description = "Daily eBay + Vinted bargain finder for Black Library books"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=22.7",
]
